// src/app/interface/user.interface.ts
import { Coordinates } from "./coordinates.interface";

export interface User{
    uuid: string; 
    firstName: string;
    lastName: string;
    email: string;
    userName: string;
    gender: string;
    address: string;
    dateOfBirth: string;
    phone : string;
    imageUrl: string;
    coordinates: Coordinates;
}